requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.3
httpx==0.28.1
selenium==4.35.0
webdriver-manager==4.0.2
markitdown[all]==0.1.3
python-dotenv==1.1.1
beautifulsoup4==4.13.4
lxml==6.0.1
openai==1.107.0
tqdm==4.67.1
